+++ R/FuzzyBetaNaiveBayes.R
@@ -71,7 +71,7 @@ FuzzyBetaNaiveBayes.default <- function(train, cl, cores = 2, fuzzy = TRUE) {
   Freq <- Freq(dados, M, Comprim_Intervalo, Sturges, minimos, cols);
   Pertinencia <- Pertinencia(Freq, dados, M);
   #------
-  # Probabilidade a priori das classes - consideradas iguais
+  # A priori probability of classes - considered equal
   pk <- rep(1 / length(unique(M)), length(unique(M)))
   #-------------------------------------------------------
 
@@ -117,7 +117,6 @@ predict.FuzzyBetaNaiveBayes <- function(object,
                                          type = "class",
                                          ...) {
   #--------------------------------------------------------
-  # type <- match.arg("class")
   test <- as.data.frame(newdata)
   #--------------------------------------------------------
   parametersC <- object$parametersC
@@ -152,7 +151,6 @@ predict.FuzzyBetaNaiveBayes <- function(object,
   #---------
   if (type == "class") {
     #-------------------------
-    #R_M_obs <- matrix(R_M_obs,nrow = N_test)
     R_M_obs <- sapply(1:nrow(R_M_obs), function(i) which.max(R_M_obs[i, ]))
     resultado <- unique(M)[R_M_obs]
     return(as.factor(c(resultado)))