+++ R/GauNBFuzzyParam.R
@@ -7,6 +7,8 @@
 #' @param cl factor of true classifications of training set
 #' @param alphacut value of the alpha-cut parameter, this value is between 0 and 1.
 #' @param metd Method of transforming the triangle into scalar, It is the type of data entry for the test sample, use metd 1 if you want to use the Yager technique, metd 2 if you want to use the Q technique of the uniformity test (article: Directional Statistics and Shape analysis), and metd 3 if you want to use the Thorani technique
+#' @param alp When metd for 4, it is necessary to have alp which are alpha-cut defined
+#' @param w When metd for 4, it is necessary to have w which are alpha-cut weights defined
 #' @param cores  how many cores of the computer do you want to use (default = 2)
 #'
 #' @return A vector of classifications
@@ -41,12 +43,12 @@
 #' @importFrom Rdpack reprompt
 #'
 #' @export
-GauNBFuzzyParam <- function(train, cl, alphacut = 0.0001, metd = 2, cores = 2) {
+GauNBFuzzyParam <- function(train, cl, alphacut = 0.0001, metd = 2, alp = c(0.35, 0.7, 0.86), w = c(0.1,0.3,0.6), cores = 2) {
   UseMethod("GauNBFuzzyParam")
 }
 
 #' @export
-GauNBFuzzyParam.default <- function(train, cl, alphacut = 0.0001, metd = 2, cores = 2) {
+GauNBFuzzyParam.default <- function(train, cl, alphacut = 0.0001, metd = 2, alp = c(0.35, 0.7, 0.86), w = c(0.1,0.3,0.6), cores = 2) {
 
   # --------------------------------------------------------
   # Estimating class parameters
@@ -65,7 +67,11 @@ GauNBFuzzyParam.default <- function(train, cl, alphacut = 0.0001, metd = 2, core
   # --------------------------------------------------------
   # --------------------------------------------------------
   # Estimating Triangular Parameters
-  alpha <- seq(alphacut, 1, 0.1)
+  alpha = alp
+  if (metd != 4) {
+    alpha <- seq(alphacut, 1.1, 0.1)
+    alpha <- ifelse(alpha > 1, 1, alpha)
+  }
   # -------------------------------
   N <- nrow(dados) # Number of observations
   # -------------------------------
@@ -118,7 +124,8 @@ GauNBFuzzyParam.default <- function(train, cl, alphacut = 0.0001, metd = 2, core
     M = M,
     alpha = alpha,
     metd = metd,
-    cores = cores
+    cores = cores,
+    w = w
   ),
   class = "GauNBFuzzyParam"
   )
@@ -144,7 +151,6 @@ predict.GauNBFuzzyParam <- function(object,
                                     type = "class",
                                     ...) {
   # --------------------------------------------------------
-  # type <- match.arg("class")
   test <- as.data.frame(newdata)
   # --------------------------------------------------------
   Parameters_varian <- object$Parameters_varian
@@ -156,6 +162,7 @@ predict.GauNBFuzzyParam <- function(object,
   alpha <- object$alpha
   metd <- object$metd
   cores <- object$cores
+  w <- object$w
   # --------------------------------------------------------
 
   # --------------------------------------------------------
@@ -176,8 +183,10 @@ predict.GauNBFuzzyParam <- function(object,
     triangulos_obs <-
       lapply(1:length(medias), function(i) { # loop to groups
         trian <- lapply(1:length(medias[[1]]), function(k) { # loop to dimensions
-          t(sapply(1:2, function(j) {
-            if(j == 2 ) j = length(alpha)
+          nn <- length(alpha)
+          if(metd != 4){ nn <- 2 }
+          t(sapply(1:nn, function(j) {
+            if((j == 2) && (metd != 4) ){ j = length(alpha) }
             # ------------
             a <- dnorm(x = as.numeric(x[k]), mean = as.numeric(Parameters_media[[i]][[k]][j, 1]), sd = sqrt(as.numeric(Parameters_varian[[i]][[k]][j, 1])))
             b <- dnorm(x = as.numeric(x[k]), mean = as.numeric(Parameters_media[[i]][[k]][j, 2]), sd = sqrt(as.numeric(Parameters_varian[[i]][[k]][j, 2])))
@@ -194,7 +203,8 @@ predict.GauNBFuzzyParam <- function(object,
       })
     # ------------
     # Center of Mass Calculation
-    vec_trian <- lapply(1:length(unique(M)), function(i) c(triangulos_obs[[i]][1, 1], triangulos_obs[[i]][2, 1], triangulos_obs[[i]][1, 2]))
+    vec_trian <- triangulos_obs
+    if(metd != 4) vec_trian <- lapply(1:length(unique(M)), function(i) c(triangulos_obs[[i]][1, 1], triangulos_obs[[i]][2, 1], triangulos_obs[[i]][1, 2]))
     # --------------------------------------------------------
     # Transforming Vector to Scalar
     # ------------
@@ -218,10 +228,14 @@ predict.GauNBFuzzyParam <- function(object,
         # Thorani Distance
         Thoranidistance(vec_trian, M)
         # ------------
+      },
+      "4" = {
+        # ------------
+        # Alpha-Order for a class of fuzzy sets
+        AlphaOrderFuzzy(vec_trian, w, M)
       }
     )
     # --------------------------------------------------------
-    # R_M_class <- which.max(produto)
     R_M_class <- R_M
     # --------------------------------------------------------
     return(R_M_class)