+++ R/PoiNBFuzzyParam.R
@@ -5,7 +5,10 @@
 #'
 #' @param train matrix or data frame of training set cases.
 #' @param cl factor of true classifications of training set
+#' @param alphacut value of the alpha-cut parameter, this value is between 0 and 1.
 #' @param metd Method of transforming the triangle into scalar, It is the type of data entry for the test sample, use metd 1 if you want to use the Yager technique, metd 2 if you want to use the Q technique of the uniformity test (article: Directional Statistics and Shape analysis), and metd 3 if you want to use the Thorani technique
+#' @param alp When metd for 4, it is necessary to have alp which are alpha-cut defined
+#' @param w When metd for 4, it is necessary to have w which are alpha-cut weights defined
 #' @param cores  how many cores of the computer do you want to use (default = 2)
 #'
 #' @return A vector of classifications
@@ -50,12 +53,12 @@
 #' @importFrom MASS fitdistr
 #'
 #' @export
-PoiNBFuzzyParam <- function(train, cl, metd = 1, cores = 2) {
+PoiNBFuzzyParam <- function(train, cl, alphacut = 0.0001, metd = 2, alp = c(0.35, 0.7, 0.86), w = c(0.1,0.3,0.6), cores = 2) {
   UseMethod("PoiNBFuzzyParam")
 }
 
 #' @export
-PoiNBFuzzyParam.default <- function(train, cl, metd = 1, cores = 2) {
+PoiNBFuzzyParam.default <- function(train, cl, alphacut = 0.0001, metd = 2, alp = c(0.35, 0.7, 0.86), w = c(0.1,0.3,0.6), cores = 2) {
 
   # --------------------------------------------------------
   # Estimating class parameters
@@ -90,14 +93,18 @@ PoiNBFuzzyParam.default <- function(train, cl, metd = 1, cores = 2) {
   lambda_estimated <- lapply(1:length(unique(M)), function(i) {
     Sub <- subset(dados, M == unique(M)[i])
     sapply(1:cols, function(j) {
-      saida <- MASS::fitdistr(Sub[, j], "poisson")[1:2]#mean(Sub[, j])
+      saida <- MASS::fitdistr(Sub[, j], "poisson")[1:2]
       return(saida)
     })
   })
   # --------------------------------------------------------
   # --------------------------------------------------------
   # Estimating Triangular Parameters
-  alpha <- seq(0.0001, 1.1, 0.1)
+  alpha = alp
+  if (metd != 4) {
+    alpha <- seq(alphacut, 1.1, 0.1)
+    alpha <- ifelse(alpha > 1, 1, alpha)
+  }
   # -------------------------------
   N <- nrow(dados) # Number of observations
   # -------------------------------
@@ -124,7 +131,8 @@ PoiNBFuzzyParam.default <- function(train, cl, metd = 1, cores = 2) {
     M = M,
     alpha = alpha,
     metd = metd,
-    cores = cores
+    cores = cores,
+    w = w
   ),
   class = "PoiNBFuzzyParam"
   )
@@ -137,8 +145,6 @@ print.PoiNBFuzzyParam <- function(x, ...) {
   # -----------------
   cat("\nFuzzy Poisson Naive Bayes Classifier for Discrete Predictors\n\n")
   # -----------------
-  #cat("Variables:\n")
-  #print(names(x$lambda_estimated[[1]]))
   cat("Class:\n")
   print(levels(x$M))
   # -----------------
@@ -150,7 +156,6 @@ predict.PoiNBFuzzyParam <- function(object,
                                     type = "class",
                                     ...) {
   # --------------------------------------------------------
-  # type <- match.arg("class")
   test <- as.data.frame(newdata)
   # --------------------------------------------------------
   Parameters_lambdas <- object$Parameters_lambdas
@@ -160,6 +165,7 @@ predict.PoiNBFuzzyParam <- function(object,
   alpha <- object$alpha
   metd <- object$metd
   cores <- object$cores
+  w <- object$w
   # --------------------------------------------------------
 
   # --------------------------------------------------------
@@ -180,7 +186,10 @@ predict.PoiNBFuzzyParam <- function(object,
     triangulos_obs <-
       lapply(1:length(unique(M)), function(i) { # loop to groups
         trian <- lapply(1:cols, function(k) { # loop to dimensions
-          t(sapply(1:length(alpha), function(j) {
+          nn <- length(alpha)
+          if(metd != 4){ nn <- 2 }
+          t(sapply(1:nn, function(j) {
+            if((j == 2) && (metd != 4) ){ j = length(alpha) }
             # ------------
             a <- dpois(x = as.numeric(x[k]), lambda =  as.numeric(Parameters_lambdas[[i]][[k]][j, 1]))
             b <- dpois(x = as.numeric(x[k]), lambda =  as.numeric(Parameters_lambdas[[i]][[k]][j, 2]))
@@ -197,7 +206,8 @@ predict.PoiNBFuzzyParam <- function(object,
       })
     # ------------
     # Center of Mass Calculation
-    vec_trian <- lapply(1:length(unique(M)), function(i) c(triangulos_obs[[i]][1, 1], triangulos_obs[[i]][11, 1], triangulos_obs[[i]][1, 2]))
+    vec_trian <- triangulos_obs
+    if(metd != 4) vec_trian <- lapply(1:length(unique(M)), function(i) c(triangulos_obs[[i]][1, 1], triangulos_obs[[i]][2, 1], triangulos_obs[[i]][1, 2]))
     # --------------------------------------------------------
     # Transforming Vector to Scalar
     # ------------
@@ -221,10 +231,14 @@ predict.PoiNBFuzzyParam <- function(object,
                     # Thorani Distance
                     Thoranidistance(vec_trian, M)
                     # ------------
+                  },
+                  "4" = {
+                    # ------------
+                    # Alpha-Order for a class of fuzzy sets
+                    AlphaOrderFuzzy(vec_trian, w, M)
                   }
     )
     # --------------------------------------------------------
-    # R_M_class <- which.max(produto)
     R_M_class <- R_M
     # --------------------------------------------------------
     return(R_M_class)