+++ NEWS.md
@@ -1,11 +1,14 @@
 # FuzzyClass (development version)
 
+# FuzzyClass 0.1.3.900
+
+- Added a new ordering method for functions with fuzzy parameters: `GauNBFuzzyParam`, `ExpNBFuzzyParam`,`GamNBFuzzyParam` and `PoiNBFuzzyParam`. The form can be used considering `metd = 4`.
 
 # FuzzyClass 0.1.2.900
 
 - Update in parameters estimation in function `FuzzyTriangularNaiveBayes`.
 - Update `predict` has been optimized to functions `FuzzyGammaNaiveBayes`.
-- Update functions of `gamNBFuzzyParam` and `GauNBFuzzyParam`.
+- Update functions of `GamNBFuzzyParam` and `GauNBFuzzyParam`.
 - `predict` function has been optimized to `FuzzyGammaNaiveBayes`, `FuzzyBetaNaiveBayes`, `FuzzyBinomialNaiveBayes`,`FuzzyExponencialNaiveBayes`,`FuzzyTriangularNaiveBayes`,`FuzzyPoissonNaiveBayes`
 
 # FuzzyClass 0.1.2